Split collection hit target widget
