+++ src/gui_app/folder_browser/collections.rs
@@ -1,18 +1,15 @@
 use std::{collections::BTreeMap, path::PathBuf};
 
 use radiant::{
-    gui::types::{Point, Rect, Rgba8},
-    layout::{LayoutOutput, Vector2},
-    runtime::{PaintFillRect, PaintPrimitive, PaintStrokeRect, PaintText, PaintTextRun},
-    theme::ThemeTokens,
-    widgets::{
-        DragHandleMessage, FocusBehavior, PaintBounds, PointerButton, TextInputMessage, TextWrap,
-        Widget, WidgetCommon, WidgetInput, WidgetOutput, WidgetSizing,
-    },
+    gui::types::{Point, Rgba8},
+    widgets::{DragHandleMessage, TextInputMessage},
 };
 use wavecrate::sample_sources::SampleCollection;
 
 use super::{FolderBrowserDrag, FolderBrowserState};
+
+mod collection_hit_target;
+pub(in crate::gui_app) use collection_hit_target::{CollectionHitMessage, CollectionHitTarget};
 
 pub(in crate::gui_app) const COLLECTION_ROW_HEIGHT: f32 = 22.0;
 pub(in crate::gui_app) const MIN_COLLECTIONS_PANEL_HEIGHT: f32 = 72.0;
@@ -56,26 +53,6 @@ pub(super) struct CollectionRenameEdit {
 pub(in crate::gui_app) struct CollectionRenameView {
     pub(in crate::gui_app) draft: String,
     pub(in crate::gui_app) input_id: u64,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub(in crate::gui_app) enum CollectionHitMessage {
-    Activate,
-    Rename,
-    Drop,
-    HoverDropTarget(Point),
-}
-
-#[derive(Clone, Debug)]
-pub(in crate::gui_app) struct CollectionHitTarget {
-    common: WidgetCommon,
-    label: String,
-    hotkey: char,
-    color: Rgba8,
-    selected: bool,
-    drop_target: bool,
-    drag_active: bool,
-    assigned_count: usize,
 }
 
 impl FolderBrowserState {
@@ -282,172 +259,6 @@ impl FolderBrowserState {
             }
         }
         counts
-    }
-}
-
-impl CollectionHitTarget {
-    pub(in crate::gui_app) fn new(collection: &SampleCollectionView) -> Self {
-        let mut common = WidgetCommon::new(
-            0,
-            WidgetSizing::fixed(Vector2::new(1.0, COLLECTION_ROW_HEIGHT)),
-        );
-        common.focus = FocusBehavior::None;
-        common.paint.bounds = PaintBounds::ClipToRect;
-        common.paint.paints_focus = false;
-        common.paint.paints_state_layers = false;
-        Self {
-            common,
-            label: collection.name.clone(),
-            hotkey: collection.hotkey,
-            color: collection.color,
-            selected: collection.selected,
-            drop_target: collection.drop_target,
-            drag_active: collection.drag_active,
-            assigned_count: collection.assigned_count,
-        }
-    }
-}
-
-impl Widget for CollectionHitTarget {
-    fn common(&self) -> &WidgetCommon {
-        &self.common
-    }
-
-    fn common_mut(&mut self) -> &mut WidgetCommon {
-        &mut self.common
-    }
-
-    fn handle_input(&mut self, bounds: Rect, input: WidgetInput) -> Option<WidgetOutput> {
-        match input {
-            WidgetInput::PointerMove { position } => {
-                self.common.state.hovered = bounds.contains(position);
-                if self.common.state.hovered && self.drag_active && !self.drop_target {
-                    return Some(WidgetOutput::typed(CollectionHitMessage::HoverDropTarget(
-                        position,
-                    )));
-                }
-                None
-            }
-            WidgetInput::PointerPress {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => {
-                self.common.state.hovered = true;
-                self.common.state.pressed = true;
-                None
-            }
-            WidgetInput::PointerRelease {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } => {
-                let activated = self.common.state.pressed && bounds.contains(position);
-                self.common.state.pressed = false;
-                self.common.state.hovered = bounds.contains(position);
-                activated.then(|| WidgetOutput::typed(CollectionHitMessage::Activate))
-            }
-            WidgetInput::PointerDoubleClick {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => {
-                self.common.state.hovered = true;
-                self.common.state.pressed = false;
-                Some(WidgetOutput::typed(CollectionHitMessage::Rename))
-            }
-            WidgetInput::PointerDrop {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => Some(WidgetOutput::typed(CollectionHitMessage::Drop)),
-            _ => None,
-        }
-    }
-
-    fn synchronize_from_previous(&mut self, previous: &dyn Widget) {
-        let Some(previous) = previous.as_any().downcast_ref::<Self>() else {
-            return;
-        };
-        self.common.state = previous.common.state;
-    }
-
-    fn accepts_pointer_move(&self) -> bool {
-        self.drag_active || self.drop_target || self.common.state.pressed
-    }
-
-    fn append_paint(
-        &self,
-        primitives: &mut Vec<PaintPrimitive>,
-        bounds: Rect,
-        _layout: &LayoutOutput,
-        theme: &ThemeTokens,
-    ) {
-        if self.selected || self.common.state.hovered || self.drop_target {
-            let fill = if self.drop_target {
-                self.color.blend_toward(theme.bg_primary, 0.72)
-            } else if self.selected {
-                theme.accent_mint.blend_toward(theme.bg_primary, 0.82)
-            } else {
-                theme.bg_secondary.blend_toward(theme.text_primary, 0.10)
-            };
-            primitives.push(PaintPrimitive::FillRect(PaintFillRect {
-                widget_id: self.common.id,
-                rect: bounds,
-                color: fill,
-            }));
-        }
-
-        if self.drop_target {
-            primitives.push(PaintPrimitive::StrokeRect(PaintStrokeRect {
-                widget_id: self.common.id,
-                rect: Rect::from_min_max(
-                    Point::new(bounds.min.x + 1.0, bounds.min.y + 1.0),
-                    Point::new(bounds.max.x - 1.0, bounds.max.y - 1.0),
-                ),
-                color: self.color,
-                width: 1.0,
-            }));
-        }
-
-        let swatch = Rect::from_min_size(
-            Point::new(bounds.min.x + 6.0, bounds.min.y + 6.0),
-            Vector2::new(10.0, 10.0),
-        );
-        primitives.push(PaintPrimitive::FillRect(PaintFillRect {
-            widget_id: self.common.id,
-            rect: swatch,
-            color: self.color,
-        }));
-
-        primitives.push(PaintPrimitive::Text(PaintTextRun {
-            widget_id: self.common.id,
-            text: PaintText::from(format!("{}  {}", self.hotkey, self.label)),
-            rect: Rect::from_min_max(
-                Point::new(bounds.min.x + 22.0, bounds.min.y),
-                Point::new(bounds.max.x - 38.0, bounds.max.y),
-            ),
-            font_size: 12.0,
-            baseline: Some((bounds.height() * 0.5 + 12.0 * 0.35).max(0.0)),
-            color: theme.text_primary,
-            align: radiant::runtime::PaintTextAlign::Left,
-            wrap: TextWrap::None,
-        }));
-        if self.assigned_count > 0 {
-            primitives.push(PaintPrimitive::Text(PaintTextRun {
-                widget_id: self.common.id,
-                text: PaintText::from(self.assigned_count.to_string()),
-                rect: Rect::from_min_max(
-                    Point::new(bounds.max.x - 34.0, bounds.min.y),
-                    Point::new(bounds.max.x - 6.0, bounds.max.y),
-                ),
-                font_size: 11.0,
-                baseline: Some((bounds.height() * 0.5 + 11.0 * 0.35).max(0.0)),
-                color: theme.text_muted,
-                align: radiant::runtime::PaintTextAlign::Right,
-                wrap: TextWrap::None,
-            }));
-        }
     }
 }
